Create photoResistor_To_ballSwitch
This file is the code for the photoresistor to detect light and the buzzer to stop beeping. The absence of light for a variable time will initiate the beeper circuit. The presence of light will deactivate the beeper circuit and close the loop by running the variable time to reset the cycles.